+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 6.6.1 (in progress)
 
+##### New Features
+* [#2644](https://github.com/oshi/oshi/pull/2644): Add serial number to Physical Memory - [@Tegrason](https://github.com/Tegrason).
+
 ##### Bug fixes / Improvements
 * [#2645](https://github.com/oshi/oshi/pull/2645): fix getOwningProcessId sometimes return -1 on 64x linux - [@yourancc](https://github.com/yourancc).